Develop

+++ inventory_ms/infra/src/database/documents/product_documents.rs
@@ -9,10 +9,10 @@ pub struct ProductDocument {
     pub title: String,
     pub subtitle: String,
     pub authors: Vec<String>,
-    pub amount_in_stock: i32,
+    pub amount_in_stock: i64,
     pub created_at: String,
     pub updated_at: String,
-    pub num_pages: i32,
+    pub num_pages: i64,
     pub tags: Vec<String>,
 }
